requirements.txt
langchain>=1.2.0
langchain-core>=1.2.0
langchain-community>=0.4.0
langchain-google-genai>=4.0.0
langchain-text-splitters>=1.0.0
faiss-cpu>=1.9.0
streamlit>=1.50.0
python-dotenv>=1.0.0
PyMuPDF>=1.25.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
rank_bm25>=0.2.2
jieba>=0.42.1
google-genai>=1.0.0
